// backend/src/models/enums/job-complement.enum.ts
export enum JobComplementEnum {
    ALMOXARIFE = "Almoxarife",
    PEDREIRO = "Pedreiro",
    AUXILIAR_ALMOXARIFADO = "Auxiliar de Almoxarifado",
    APONTADOR = "Apontador",
    ELETRICISTA = "Eletricista",
    ENCANADOR = "Encanador",
    VIGIA = "Vigia",
    MOTOBOY_I = "Motoboy I",
    MOTOBOY_II = "Motoboy II",
    MOTOBOY_III = "Motoboy III",
    OPERADOR_TRATOR_I = "Operador de Trator I",
    OPERADOR_TRATOR_II = "Operador de Trator II",
    OPERADOR_TRATOR_III = "Operador de Trator III",
    ASSISTENTE_ADMIN_JUNIOR = "Assistente Administrativo Junior",
    ASSISTENTE_ADMIN_PLENO = "Assistente Administrativo Pleno",
    ASSISTENTE_ADMIN_SENIOR = "Assistente Administrativo Senior",
    DESENVOLVEDOR_JUNIOR = "Desenvolvedor Junior",
    DESENVOLVEDOR_PLENO = "Desenvolvedor Pleno",
    DESENVOLVEDOR_SENIOR = "Desenvolvedor Senior",
    ENCARREGADO = "Encarregado",
    MEIO_OFICIAL = "Meio-Oficial",
    AUXILIAR = "Auxiliar",
    AJUDANTE = "Ajudante",
    PROFISSIONAL = "Profissional",
    SERRALHEIRO = "Serralheiro"
}